pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code_analyzer"
version = "0.1.0"
description = "A multi-agent system that analyzes Python projects, indexes code, generates and runs pytest test cases, and produces documentation."
authors = [
    {name = "Bhanu Mahesh", email = "user@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "crewai>=0.28.8",
    "crewai_tools>=0.1.6",
    "langchain>=0.0.335",
    "langchain_openai>=0.0.2",
    "pydantic>=2.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "asyncio>=3.4.3",
    "openai>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "mypy",
]

[project.scripts]
code-analyzer = "code_analyzer.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true